+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Changed
 
+- Changed the naming convention for Fortran Objects
+  ([#109](https://github.com/gnikit/fortls/issues/109))
 - Formatted all files with `pre-commit`
 
 ## 2.3.1